--- README
@@ -1,48 +1,0 @@
-                              )                 
-                           ( /(          (  (   
-                            )\())(   (    )\))(  
-                           ((_)\ )\  )\ )((_))\  
-                           | |(_|(_)_(_/( (()(_) 
-                           | / / _ \ ' \)) _` |  
-                           |_\_\___/_||_|\__, |  
-                         secure web node |___/ v0.1.0
-
-Documentation
-=============
-  The documentation can be found in the `doc/` directory, it is
-  written in Markdown (.md) format.
-
-  - [doc/README.md]: Kong
-    
-Components (krates)
-===================
-
-  Kong consists of severel components known as krates.
-
-  - [krates/kdata]: All the common data types used in kong
-  - [krates/kerror]: Error types
-  - [krates/klient]: Clients in Javascript and Dart
-  - [krates/kogs]: Addons/Plugins
-  - [krates/kollection]: Database management
-  - [krates/konfig]: Node configuration
-  - [krates/kong]: Core server runtime
-  - [krates/kontainer]: LXC container management
-  - [krates/krypto]: Cryptography and security related code
-
-Unlicense
-=========
-
-Written and placed in the public domain by Jackson G. Kaindume.
-
-Anyone is free to copy, modify, publish, use, compile, sell, or
-distribute this software, either in source code form or as a compiled
-binary, for any purpose, commercial or non-commercial, and by any
-means.
-
-For more information, please refer to the __UNLICENSE__ file in the
-repository.
-
-
-----------------------------------------------------------------------
-            Jackson G. Kaindume <cy6erlion@protonmail.com>
-                  https://kwatafana.github.io/www/
