--- doc/README.md
@@ -1,58 +1,0 @@
----
-title: ⬡ Kong
-subtitle: Secure Web Node
-author: Jackson G. Kaindume <cy6erlion@protonmail.com>
-date: 2020-2023
-...
----
-
-
-``` text
-                              )                 
-                           ( /(          (  (   
-                            )\())(   (    )\))(  
-                           ((_)\ )\  )\ )((_))\  
-                           | |(_|(_)_(_/( (()(_) 
-                           | / / _ \ ' \)) _` |  
-                           |_\_\___/_||_|\__, |  
-                         secure web node |___/ v0.1.0
-```
-
-
-## Features
-
-- [ ] Confidentiality
-- [ ] Integrity
-- [ ] Availability
-- [ ] Database Management
-- [ ] Session Management
-- [ ] HTTP+JSON API
-- [ ] Optional Federation
-- [ ] Built in HTTP server
-- [ ] Account Management
-
-## User Manuals
-
-`kong` is made up of different components known as `krates`.
-Bellow are links to documentation of those respective `krates`:
-
-[📄 kdata](kdata.html): Data types and structures
-
-[⚠️ kerror](#kerror): Error types
-
-[📱 klient](#klient): Clients Javascript and Dart
-
-[🗄️ kollection](kollection.html): database management
-
-[🎚️ konfig](#konfig): Node configuration
-
-[⬡ kong](#kong): Core server runtime
-
-[🥷 konja](#konja): Pen-testing tools
-
-[📦 kontainer](#kontainer): LXC container management
-
-[🔐 krypto](krypto.html): Cryptography and Security
-
----
-[[⧉]](https://kwatafana.codeberg.page/)
