--- doc/kollection.md
@@ -1,22 +1,0 @@
----
-title: 🗄️ Kong Database
-author: Jackson G. Kaindume <cy6erlion@protonmail.com>
-date: Last update 28.02.2023
-...
----
-
-# 🗄️
-
-Documentation of the `kong` database management system.
-
-Kong stores structured data in a database. [SQLite](https://sqlite.org/)
-is the database management system that is used, with help from the 
-[rusqlite](https://sqlite.org/) crate.
-
-The following data is stored the database:
-
-1. [`Accounts`](./data.html#account)
-
----
-
-[[⬡]](./index.html) | [[⧉]](https://kwatafana.codeberg.page/)
