+++ krates/kdata/README.md
@@ -1,0 +1,5 @@
+---
+title: kdata
+subtitle: Common data structures used in `kong`
+...
+---
