+++ krates/kerror/README.md
@@ -1,0 +1,5 @@
+---
+title: kerror
+subtitle: Common error types used in `kong`
+...
+---
