+++ krates/klient/README.md
@@ -1,1 +1,6 @@
-# kclient
+---
+title: klient
+subtitle: `kong` HTTP clients in Javascript and Dart
+...
+---
+
