+++ krates/kogs/README.md
@@ -1,0 +1,5 @@
+---
+title: kogs
+subtitle: kong standard plugins
+...
+---
