+++ krates/kollection/src/lib.rs
@@ -7,6 +7,10 @@
 //! The following data is stored the database:
 //! - `kdata::accounts::Account`
 
+#![doc(html_favicon_url = "https://kwatafana.org/logo.jpeg")]
+#![doc(html_logo_url = "https://kwatafana.org/logo.jpeg")]
+#![warn(missing_docs, unreachable_pub, future_incompatible, rust_2018_idioms)]
+
 use kdata::accounts::{Account, PublicAccount};
 use kerror::KError;
