+++ krates/konfig/README.md
@@ -1,0 +1,5 @@
+---
+title: konfig
+subtitle: `kong` server configuration
+...
+---
