+++ krates/konfig/src/lib.rs
@@ -1,3 +1,7 @@
+#![doc(html_favicon_url = "https://kwatafana.org/logo.jpeg")]
+#![doc(html_logo_url = "https://kwatafana.org/logo.jpeg")]
+#![warn(missing_docs, unreachable_pub, future_incompatible, rust_2018_idioms)]
+
 //! server configuration
 use kerror::KError;
 use serde::Deserialize;
