--- krates/kontainer/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "kontainer"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-serde.workspace = true
-toml.workspace = true
