--- krates/kontainer/README.md
@@ -1,1 +1,0 @@
-# kontainer
