+++ krates/krypto/README.md
@@ -1,0 +1,5 @@
+---
+title: krypto
+subtitle: `kong` Cryptography and security related code
+...
+---
