+++ krates/krypto/src/lib.rs
@@ -18,6 +18,10 @@
 //!   - Encryption at Rest
 //!     - Deletion of data
 
+#![doc(html_favicon_url = "https://kwatafana.org/logo.jpeg")]
+#![doc(html_logo_url = "https://kwatafana.org/logo.jpeg")]
+#![warn(missing_docs, unreachable_pub, future_incompatible, rust_2018_idioms)]
+
 pub mod authentication;
 pub mod defaults;
 mod error;
